Refactort the code in layers
