--- server/db/mongoose.js
@@ -1,8 +1,0 @@
-const mongoose = require('mongoose');
-
-function connectToMongoDB() {
-  mongoose.Promise = global.Promise;
-  mongoose.connect(process.env.MONGODB_URI);
-}
-
-module.exports = { connectToMongoDB };
