+++ server/services/todosService.js
@@ -1,6 +1,4 @@
-const { ObjectID } = require('mongodb');
 const { Todo } = require('../models/todo');
-const _ = require('lodash');
 
 function findTodos() {
   return new Promise((resolve, reject) => {
@@ -18,10 +16,7 @@ function findTodos() {
   });
 }
 
-function saveTodo(todoText) {
-  const todo = new Todo({
-    text: todoText
-  });
+function saveTodo(todo) {
   return new Promise((resolve, reject) => {
     todo.save().then((doc) => {
       resolve({
@@ -39,12 +34,6 @@ function saveTodo(todoText) {
 
 function findTodo(id) {
   return new Promise((resolve, reject) => {
-    if (!ObjectID.isValid(id)) {
-      return reject({
-        code: 400,
-        message: 'ID not valid'
-      });
-    }
     Todo.findById(id).then((todo) => {
       if (!todo) {
         return reject({
@@ -65,13 +54,6 @@ function findTodo(id) {
 
 function removeTodo(id) {
   return new Promise((resolve, reject) => {
-    if (!ObjectID.isValid(id)) {
-      return reject({
-        code: 400,
-        message: 'ID not valid'
-      });
-    }
-
     Todo.findByIdAndRemove(id).then((todo) => {
       if (!todo) {
         return reject({
@@ -92,20 +74,6 @@ function removeTodo(id) {
 
 function updateTodo(id, body) {
   return new Promise((resolve, reject) => {
-    if (!ObjectID.isValid(id)) {
-      return reject({
-        code: 400,
-        message: 'ID not valid'
-      });
-    }
-
-    if (_.isBoolean(body.completed) && body.completed) {
-      body.completedAt = new Date().getTime();
-    } else {
-      body.completed = false;
-      body.completedAt = null;
-    }
-
     Todo.findByIdAndUpdate(id, { $set: body }, { new: true }).then((todo) => {
       if (!todo) {
         return reject({
